* 'feature/userinfo_update' of https://github.com/zhiyicx/thinksns-plus-android:
  （ #557）  code   update  recieved  comments

# Conflicts:
#	baseproject/src/main/java/com/zhiyicx/baseproject/config/ApiConfig.java

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/home/message/messagecomment/MessageCommentAdapter.java
@@ -72,7 +72,7 @@ public class MessageCommentAdapter extends CommonAdapter<CommentedBean> {
             holder.setVisible(R.id.tv_deatil, View.GONE);
             holder.setVisible(R.id.iv_detail_image, View.VISIBLE);
             mImageLoader.loadImage(getContext(), GlideImageConfig.builder()
-                    .url(ImageUtils.imagePathConvertV2(commentedBean.getTarget_image()
+                    .url(ImageUtils.imagePathConvertV2(commentedBean.getTarget_image().intValue()
                             , mContext.getResources().getDimensionPixelOffset(R.dimen.headpic_for_user_center)
                             , mContext.getResources().getDimensionPixelOffset(R.dimen.headpic_for_user_center)
                             , ImageZipConfig.IMAGE_50_ZIP))
